flytt generering av utbetalinger til egen service
