+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/utbetaling/task/OppdaterUtbetalingBeregning.kt
@@ -4,7 +4,7 @@ import com.github.kagkarlsson.scheduler.task.helper.OneTimeTask
 import com.github.kagkarlsson.scheduler.task.helper.Tasks
 import kotlinx.serialization.Serializable
 import kotliquery.Session
-import no.nav.mulighetsrommet.api.utbetaling.UtbetalingService
+import no.nav.mulighetsrommet.api.utbetaling.GenererUtbetalingService
 import no.nav.mulighetsrommet.serializers.UUIDSerializer
 import no.nav.mulighetsrommet.tasks.executeSuspend
 import no.nav.mulighetsrommet.tasks.transactionalSchedulerClient
@@ -12,7 +12,7 @@ import java.time.Instant
 import java.util.*
 
 class OppdaterUtbetalingBeregning(
-    private val utbetalingService: UtbetalingService,
+    private val utbetalinger: GenererUtbetalingService,
 ) {
 
     @Serializable
@@ -24,7 +24,7 @@ class OppdaterUtbetalingBeregning(
     val task: OneTimeTask<TaskData> = Tasks
         .oneTime(javaClass.simpleName, TaskData::class.java)
         .executeSuspend { instance, _ ->
-            utbetalingService.oppdaterUtbetalingBeregningForGjennomforing(instance.data.gjennomforingId)
+            utbetalinger.oppdaterUtbetalingBeregningForGjennomforing(instance.data.gjennomforingId)
         }
 
     fun schedule(gjennomforingId: UUID, startTime: Instant, session: Session) {
